+++ server/camcops_server/cc_modules/cc_redcap.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env python
 
-"""
-camcops_server/cc_modules/cc_redcap.py
+"""camcops_server/cc_modules/cc_redcap.py
 
 ===============================================================================
 
@@ -26,21 +25,16 @@ camcops_server/cc_modules/cc_redcap.py
 
 **Implements communication with REDCap.**
 
-(Thoughts from 2019-01-27, RNC.)
-
 - For general information about REDCap, see https://www.project-redcap.org/.
 
-- The API documentation seems not to be provided there, but is available from
+- The API documentation is not provided there, but is available from
   your local REDCap server. Pick a project. Choose "API" from the left-hand
   menu. Follow the "REDCap API documentation" link.
 
-- In Python, we have PyCap (https://pycap.readthedocs.io/ or
+- We use PyCap (https://pycap.readthedocs.io/ or
   https://github.com/redcap-tools/PyCap). See also
-  http://redcap-tools.github.io/projects/. This is no longer being actively
-  developed.
-
-- There are also Python examples in the "API Examples" section of the API
-  documentation. See, for example, ``import_records.py``.
+  http://redcap-tools.github.io/projects/. PyCap is no longer being actively
+  developed though the author is still responding to issues and pull requests.
 
 *REDCap concepts*
 
@@ -57,101 +51,18 @@ camcops_server/cc_modules/cc_redcap.py
 
 - **REDCap Shared Library:** a collection of public instruments.
 
-*My exploration*
+- **Record:** for our purposes a record corresponds to a patient. REDCap has no
+  internal concept of a patient. A record will have multiple instruments, some
+  of which may be "repeating", allowing the same form to be submitted at
+  different times.
 
-- A "record" has lots of "instruments". The concept seems to be a "study
-  visit". If you add three instruments to your project (e.g. a PHQ-9 from the
-  Shared Library plus a couple of made-up things) then it will allow you to
-  have all three instruments for Record 1.
+We use an XML fieldmap to describe how the rows in CamCOPS task tables are
+translated into REDCap records. See :ref:`<redcap>`.
 
-- Each instrument can be marked complete/incomplete/unverified etc. There's a
-  Record Status Dashboard showing these by record ID. Record ID is an integer,
-  and its field name is ``record_id``. This is the first variable in the data
-  dictionary.
-
-- The standard PHQ-9 (at least, the most popular in the Shared Library) doesn't
-  autocalculate its score ("Enter Total Score:")...
-
-- If you import a task from the Shared Library twice, you get random fieldnames
-  like this (see ``patient_health_questionnaire_9b``):
-
-  .. code-block:: none
-
-    Variable / Field Name	    Form Name
-    record_id	                my_first_instrument
-    name	                    my_first_instrument
-    age	                        my_first_instrument
-    ipsum	                    my_first_instrument
-    v1	                        my_first_instrument
-    v2	                        my_first_instrument
-    v3	                        my_first_instrument
-    v4	                        my_first_instrument
-    phq9_date_enrolled	        patient_health_questionnaire_9
-    phq9_first_name	            patient_health_questionnaire_9
-    phq9_last_name	            patient_health_questionnaire_9
-    phq9_1	                    patient_health_questionnaire_9
-    phq9_2	                    patient_health_questionnaire_9
-    phq9_3	                    patient_health_questionnaire_9
-    phq9_4	                    patient_health_questionnaire_9
-    phq9_5	                    patient_health_questionnaire_9
-    phq9_6	                    patient_health_questionnaire_9
-    phq9_7	                    patient_health_questionnaire_9
-    phq9_8	                    patient_health_questionnaire_9
-    phq9_9	                    patient_health_questionnaire_9
-    phq9_total_score	        patient_health_questionnaire_9
-    phq9_how_difficult	        patient_health_questionnaire_9
-    phq9_date_enrolled_cdda47	patient_health_questionnaire_9b
-    phq9_first_name_e31fec	    patient_health_questionnaire_9b
-    phq9_last_name_cf0517	    patient_health_questionnaire_9b
-    phq9_1_911f02	            patient_health_questionnaire_9b
-    phq9_2_258760	            patient_health_questionnaire_9b
-    phq9_3_931d98	            patient_health_questionnaire_9b
-    phq9_4_8aa17a	            patient_health_questionnaire_9b
-    phq9_5_efc4eb	            patient_health_questionnaire_9b
-    phq9_6_7dc2c4	            patient_health_questionnaire_9b
-    phq9_7_90821d	            patient_health_questionnaire_9b
-    phq9_8_1e8954	            patient_health_questionnaire_9b
-    phq9_9_9b8700	            patient_health_questionnaire_9b
-    phq9_total_score_721d17	    patient_health_questionnaire_9b
-    phq9_how_difficult_7c7fbd	patient_health_questionnaire_9b
-
-*The REDCap API*
-
-- The basic access method is a URL for a server/project plus a project-specific
-  security token.
-
-- Note that the API allows you to download the data dictionary.
-
-*Other summaries*
-
-- https://github.com/nutterb/redcapAPI/wiki/Importing-Data-to-REDCap is good.
-
-*So, for an arbitrary CamCOPS-to-REDCap mapping, we'd need:*
-
-#.  An export type of "redcap" with a definition including a URL and a project
-    token.
-
-#.  A configurable patient ID mapping, e.g. mapping CamCOPS forename to a
-    REDCap field named ``forename``, CamCOPS ID number 7 to REDCap field
-    ``my_study_id``, etc.
-
-#.  Across all tasks, a configurable CamCOPS-to-REDCap field mapping
-    (potentially incorporating value translation).
-
-    - A specimen translation could contain the "default" instrument fieldnames,
-      e.g. "phq9_1" etc. as above.
-
-    - This mapping file should be separate from the patient ID mapping, as the
-      user is quite likely to want to reuse the task mapping and alter the
-      patient ID mapping for a different study.
-
-    - UNCLEAR: how REDCap will cope with structured sub-data for tasks.
-
-#.  A method for batching multiple CamCOPS tasks into the same REDCap record,
-    e.g. "same day" (configurable?), for new uploads.
-
-#.  Perhaps more tricky: a method for retrieving a matching record to add a
-    new task to it.
+REDCap does not assign instance IDs for repeating instruments so we need to
+query the database in order to determine the next instance ID. It is possible
+to create a race condition if more than one client is trying to update the same
+record at the same time.
 
 """