+++ docs/source/changelog.rst
@@ -2733,3 +2733,5 @@ Current C++/SQLite client, Python/SQLAlchemy server
 
 - ``cardinal_pythonlib==1.0.81`` for memory efficiency, then for the
   :func:`nice_call` function that sorts this out a bit, etc.
+
+- :ref:`Export of tasks to REDCap <redcap>`.